Implements basis for handling FHRs with a particular image format
While the final implementation should handle Custom Resource
manifests in an easily extendable way, the solution is more
complex. To be able to start testing helm integration, an approach
to provide quick and dirty solution was asked for.
This solution only caters for FHR kind and the image format provided
as Spec.Values.Image (eg quay.io/weaveworks/server-io/master-a00001)

The tryUpdate method still to be provided.